Load descs on first render
This is such a stupid fix. We need to setState so that the tile element re-renders properly BUT it's impossible to make a dynamically keyed object in states unless I'm like MEGA stupid or something. So the fix is to just make a state that never gets used for anything to trigger re-render after setting the map.